MoneyMaker.
Money Maker is a simple projects for converting cents into coins, this program can be manipulated and fixed in many ways to fit many situations. Also it taught me personally a lot about variables and different data types and how to manage them.
This simple code can be convenient for places that use coins for playing games or winning prizes, or at least that was what I was thinking while developing it.